+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [1.1.0] / 2025-09-21
 
 ### Added
 - Expose frontmatter to JavaScript Macro
@@ -15,6 +15,10 @@ See:[Issue #2](https://github.com/CePeU/MarkdownToFoundry/issues/2)
 - Version information to be shown to the user
 - Version information has been added to the profiles
 
+- Make debug output optional
+A new setting was implemented to select debug output
+See: [Issue #13](https://github.com/CePeU/MarkdownToFoundry/issues/13)
+
 ### Fixed
 - SVG Macro Icon is too dark/not visible on Foundry VTT 12.xxx
 Reworked the SVG for the macro icon to be a bit smaller and have white stroke color
@@ -25,7 +29,9 @@ See:[Issue #16](https://github.com/CePeU/MarkdownToFoundry/issues/16)
 All linking/relinking selections have been put under  **Foundry journal relinking export settings**
 - The plugin is now able to check the version information and reimport profiles to new profile structures
 - Readme was adjusted to include documentation for new macro settings
-
+- **Changed picture collection function buildPictureUploadList to work on rendered HTML instead of embedds.**
+This allows for Dataview and other plugins to render completly and including pictures and to export the rendered result.
+This will also make it easier to export the rendered HTML to file paths and adjust the HTML for relative file paths in the filesystem.
 
 ## [1.0.2] / 2025-09-14
 ### Removed